+++ mobile/src/app/(tabs)/_layout.tsx
@@ -35,20 +35,12 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="map"
+        name="search"
         options={{
-          title: 'Map',
-          tabBarLabel: 'Map',
+          title: 'Search',
+          tabBarLabel: 'Search',
           headerShown: false, // Add this line to hide the header
-          tabBarIcon: ({ color }: Color) => <MaterialIcons name="map" size={28} color={color} />,
-        }}
-      />
-      <Tabs.Screen
-        name="showcase"
-        options={{
-          title: 'Showcase',
-          tabBarLabel: 'Showcase',
-          tabBarIcon: ({ color }: Color) => <MaterialIcons name="code" size={28} color={color} />,
+          tabBarIcon: ({ color }: Color) => <MaterialIcons name="search" size={28} color={color} />,
         }}
       />
     </Tabs>
